change: use double to fix total cost

+++ src/main/java/cz/cvut/fel/aic/simod/ridesharing/insertionheuristic/InsertionHeuristicSolver.java
@@ -535,7 +535,7 @@ public class InsertionHeuristicSolver<T> extends DARPSolver implements EventHand
 		}
 
 		// cost computation
-		double newPlanCost = planCostProvider.calculatePlanCost(newPlanDiscomfort, newPlanTravelTime / 1000);
+		double newPlanCost = planCostProvider.calculatePlanCost(newPlanDiscomfort, (double) newPlanTravelTime / 1000);
 
 		return new DriverPlan(newPlanTasks, newPlanTravelTime / 1000, newPlanCost);
 	}
